Tell ack to ignore doc directory
I'm not totally sure about this. Most of the time, this is *probably*
what you want, but there are times when you need to include the docs
dir. I currently fall back to using grep in a terminal when I want to
search more than just code source files, but that's just me ;-)