Bring back vim mru plugin
CtrlPMRU is great, but it's often preferable to load the list of MRU
files in the current window rather than a new window. MRU files plugin
supports this, so bring it back, and configure to use current window.

Note: now using fork of MRU that plays nicely with alternate file :-)