Stop a simple brew install randomly breaking shit
I know, I know, auto-update is a good thing, except it's not when a
simple brew install updates some core dependencies like python and
then breaks programs and scripts that have not been tested with the
latest stable version, especially when you rely on those programs and
scripts both to do your own work, and help others to do theirs.

The example here that drove me up the fscking wall was a simple install
of aws-rotate-iam-keys, which is just a shell script, but it took an age
for some people to install as it updated python from 3.6 to 3.7, and it
then broke pylint, which prevented people from pushing changes to python
repos without upgrading pylint and addressing new warning in pylint v2.