Hack vim long line highlighting for editorconfig
I don't use the max line indicator option from the editorconfig plugin,
because I'm not a fan of any of the available options, but the plugin
changes the textwidth of a buffer, so I need to make sure my own long
line highlighting is applied after the plugin changes the textwidth.

A generic solution to this is to call a function on VimEnter that adds
the autocmd and also highlights the long lines for existing buffers.
There are other options, including hooks provided by the editorconfig
plugin itself, but I thought the generic VimEnter option was simplest.