Don't override default vim formatoptions anymore
No longer necessary, was originally added as a few separate additions to
formatoptions to make things work nicely with editorconfig, but without
editorconfig overriding formatoptions and enabling auto-wrapping. This
doesn't seem necessary anymore, the ftplugins for all the languages I
use regularly set formatoptions to something sensible for the language.