Fix git config to always fetch tags for git 2.21+
Adding tags to the refspec now causes problems cloning repos, e.g.
fatal: multiple updates for ref 'refs/tags/v1.0.0' not allowed

This is apparently a new issue since git version 2.21.0, but this
refspec is no longer required to force git to always fetch tags.

It looks like just setting the tagOpt will do it nowadays, and if not
there is always the option of aliasing fetch to fetch --tags, as the
--tags option now fetches tags in "addition" to everything else.