Fix git mru alias to work with newly cloned repos
The initial entry in the reflog is a clone, not a checkout or rename,
so mru alias wasn't working well for newly cloned repos, e.g. clone
repo, create new branch, and git mru would not show the main branch.

Fix by always prepending the symbolic ref to the last checkout to the
list of mru branches, it's a bit hacky but it does the job well enough.