Simplify vim supertab configuration, no magic
Disable context aware completion, because vim omnicompletion doesn't
work very well out of the box, and is especially poor with javascript,
which I'm spending more and more time with recently (for my sins).

This is a solvable problem, there are lots of great completion plugins,
but as a default starting point, using tab for keyword completion is ok.

Arguably this change means supertab is no longer necessary, may remove.