Add vim mapping to expand/restore current window
An alternative to using ZoomWin, and possibly a simple replacement.

While it's not quite the same as it leaves all windows open, with all
but the current window shrunk to one row or column, it's also very fast
and reliable as it lacks the ZoomWin magic to hide and restore windows.