Use QFEnter vim plugin to tame qf open behaviour
Vim by default always skips past special buffers when trying to open
files from the quickfix window, which is a sensible default as it
doesn't clobber nofile buffers loaded by plugins etc., but there are
often exceptions to this simple rule, for example, I open bufexplorer
and mru buffers in the current window rather than a split, and I want
to reuse those buffers when opening files from quickfix windows.

The QFEnter plugin solves this problem by opening files in the previous
window, and allowing some filetypes to be excluded, e.g. nerdtree, but
it just stops with an error by default when the previous window matches
an excluded filetype, so I've added a custom CR mapping to find the next
usable window rather than just error when hitting an excluded filetype.

This even works when there are no usable windows, i.e. all windows are
either excluded filetypes or the quickfix window itself, QFEnter seems
to fall back to Vim's default behaviour of opening a split in this case.