Support ripgrep as grepprg in vim
If ag executable not found, use rg as grepprg. I personally prefer the
default behaviour of ag with nested repos, it does not use the parent
ignore, but will still use global ignores (including a global .agignore)

While rg can be configured to do the same, I find the default behaviour
a little odd (probably because I'm used to ag, rather than it being bad)

Arguably this is a pointless commit at the moment, but I'm thinking
along the lines of being able to easily switch between grep programs.

Why even bother with that? I work with lots of people who use vscode,
vscode uses ripgrep with a default custom config out of the box, I want
to be able to easily replicate that in vim so I can see what they see.