Add custom grep prompt to vim, inspired by grepper
Make is easier to search in files using grepprg by adding an input
prompt which automatically escapes vim cmdline special characters.

As per previous custom grep command, also runs grep silently to avoid
the "Press ENTER to continue" prompt after running an external command.

This was inspired by vim-grepper, which is a great plugin, but seems
overkill for my simple needs - I don't need to support multiple grep
programs, or even toggle between different sets of options for a single
grep program, and I don't want to run my grep program asynchronously (if
it's running slowly, I want to see that, hit CTRL-C, and then fix it).
All I want is the prompt, which is a brilliant idea, so I've reproduced
a minimalist version of that feature with slightly improved completion.