Disable some macvim hacks, no longer necessary?
On previous versions of macvim there was a glitch whereby if vim lost
focus while a bottom split (may have only been botright) was active, it
resized slightly when focus was regained, resulting in unequal windows.

This seems to be fixed, at least for focus lost and gained. Of course,
it could be something else that changed on this laptop which was dusted
off recently, but assume it is fixed in macim and comment out for now.