Use MruGetFiles() with FZFRecent vim command
This command loads FZF with MRU files in the current working directory.
Make sure it loads the same files as the <leader>r mapping to list MRU
files in the current working directory, by using MruGetFiles() function.

Note: this is actually slower than calling the external grep command,
but not enough for it to allow it to behave differently (the MRU code to
get the list of files changes dependent on the value of MRU_FuzzyMatch).