Improve git mru alias for newly cloned repos
mru alias relies on parsing reflog looking for checked out or renamed
branches, which doesn't work for newly cloned repos. Previous hack to
work around this just prepended the symbolic ref to the last checkout
to the list of mru branches, i.e. HEAD@{-1}, but that only works when
HEAD@{-1} has a symbolic ref, which is not guaranteed, and for example
fails after cloning a repo and then checking out a few sha1 refs, e.g.

    git clone url-for-repo-foo
    cd foo
    git checkout e0789536
    git checkout 60f03d6c
    git mru -> empty

Fix by checking if parsing reflog for checked out or renamed branches
resulted in an empty string, and if so just print the local version of
the symbolic ref pointed to by refs/remotes/origin/HEAD, which is what
we want on a freshly cloned repo (without having to parse reflog again).
Though the ref is a "remote", this still works for locally cloned repos.