Improve git mru alias again for newly cloned repos
The workaround for dealing with a fresh clone followed by a number of
detached head checkouts introduced a bug whereby with a fresh clone
followed by a new branch the mru list would not include the main branch,
because that branch would not show up as a checked out or moved branch
in the reflog, but was also no longer being prepended to the mru list.

Fix by always appending the main branch to the list of mru branches.