Remove vim WhichKey g mapping - breaks g, command
My leader key is mapped to comma, and easyclip plugin adds g<leader>
mappings, which combined with WhichKey mapping for g key, breaks built
in "g," command which moves cursor to the next position in change list.

I could probably fix this with a little hacking, but I really don't
need the which key menu for g key mappings anyway, they are few.