Fix wincmd p when re-using existing vim qf window
Running a quickfix command when the quickfix window is the active window
resets the previous window to the first window, which means my q mapping
to quit/close and go to previous window doesn't always work as expected.

Use QuickFixCmdPre autocmd to hack a workaround for now. I did try
tracking the opener in a window local variable, but they seem to get
wiped out every time you run a new quickfix command, and a buffer local
variable, but these persist unless the buffer is wiped out when closed.