Remember cursor position with visual modes in vim
Set previous context mark when entering visual modes in vim, and return
to mark when exiting visual modes and returning back to normal mode.

This is mostly for when I use the wrong text object to select something
and want to cancel quickly and start again with a different text object.