Restore old NERDTree/Newrw leader mappings, d/e
I never got used to _ to toggle netrw, and find the old "d" for project
drawer and "e" for split explorer mnemonics just make a lot of sense.

Using i for vimspector to "inspect" rather than "debug" is a bit of a
stretch, but it does nicely tie in with chrome/safari devtools, and I'm,
for now anyway, going to avoid using capital letters in leader mappings.