Tweak whitepsace highlighting in vim again
I had a hunch that setting a colour outside the basic 16 colours might
come back to bite me and sure enough it has, colour number 16 on another
macbook with the same term program and colour scheme but a different
version of MacOS renders colour number 16 differently. Not sure why and
don't have time to investigate, so play safe and stay within 16 colours.

Note: this is the same colour number used for comments in solarized, so
I'm pretty confident it'll render the same on both macbooks/terminals.

Also remove bold styling in both terminal and gui, unnecessarily loud.