Revert "Use vim's built in package support for plugins"
Using vim's build in package support makes it a little painful to work
with different vim configurations for different editing contexts, e.g.
a vimrc for general editing, and another minimal vim for commit messages

My use case here is to develop a minimal, but friendly, vimrc to share
with colleagues, and I want to eat my own dog food. While I could set
packpath=$VIMRUNTIME in the minimal vimrc, that just over-complicates
things and also makes it hard to move the minimal vimrc to a plugin.

This reverts commit 8fbe0611413c3ce2a460a65fe25ee1c1fbc79876.