Set gui vim colors to work with "open" command
When the open command on MacOS associates a file with MacVim it seems
to send the terminal environment to MacVim, which makes Vim set the
background based on the terminal background, and breaks my code to set
the colorscheme correctly as the $TERM_PROGRAM env var is set to iTerm.

Fix by checking for gui first, and forcing gui background to light.

Note: I guess this probably affect xdg-open too, but have not tested.