Simplify vim visual mode restore cursor pos hack
This was only really added for text objects, because after visually
selecting using a text object, then realising you've not quite got it
right, you really want to return to the previous cursor position, but
the implementation was overly complicated because it dealt with visual
mode in general, including trying to handle selection using the mouse,
leading to edge cases that were not handled, and some weird behaviour.

Simplify by just remapping vi and va in normal mode and using a hard
coded named mark. It's crude, but it works, without many side effects.