Simplify vim visual mode restore cursor pos hack
This was only really added for text objects, because after visually
selecting using a text object, then realising you've not quite got it
right, you really want to return to the previous cursor position, but
the implementation was overly complicated because it used a ModeChanged
event to avoid mapping the escape key. This resulted in some weird
behaviour when changing text that was visually selected using a motion
or mouse movement that went backwards (cursor seemed to jumped forward).

Fix by just mapping ESC to return to the previous cursor position. Works
fine in iTerm2, Apple Terminal and MacVim, which means it works for me.

This also does not mess with other methods of exiting visual mode, so
you can leave visual mode and retain the new cursor position if needed.