Improve vim-cutlass mappings, m and M alternatives
vim-cutlass changes delete operations to just delete, and not yank, so
I've added "move" mappings to delete and yank (as recommended). These
mappings break built-in commands to set named marks and move to the
middle of the screen, so add alternative mappings for those commands.

Note: Z chosen to jump to middle of screen as it corresponds nicely with
zz to move the cursor line to the middle of the screen. I might want to
disable ZZ though, in case I hit it by accident and write+close file.