Add .node-version file for vimspector adapters
vscode-node-debug2 adapter requires node 16 or lower, running npm run
build on later versions generates the following error:

[15:11:06] Error: You must provide the URL of lib/mappings.wasm by calling SourceMapConsumer.initialize({ 'lib/mappings.wasm': ... }) before using SourceMapConsumer
    at readWasm (/Users/mwoods/dotfiles/vim/pack/bundle/start/vimspector/gadgets/macos/download/vscode-node-debug2/node_modules/gulp-typescript/node_modules/source-map/lib/read-wasm.js:8:13)
    at wasm (/Users/mwoods/dotfiles/vim/pack/bundle/start/vimspector/gadgets/macos/download/vscode-node-debug2/node_modules/gulp-typescript/node_modules/source-map/lib/wasm.js:25:16)
    at /Users/mwoods/dotfiles/vim/pack/bundle/start/vimspector/gadgets/macos/download/vscode-node-debug2/node_modules/gulp-typescript/node_modules/source-map/lib/source-map-consumer.js:264:14
    at process.processTicksAndRejections (node:internal/process/task_queues:95:5)

This is apparently a known issue, and I do vaguely recall working around
it in the past before re-installing plugins, so just add a .node-version
file with a version supported by vscode-node-debug2 and make shit work.

This is a horrible hack, but I've just lost 30 minutes trying to figure
out WTF was going on, and I don't want to repeat that in a few months.