Replace vim lexima with auto-pairs and endwise
I've been running with this locally for a while, without realising I
hadn't committed the changes, and the combination of auto-pairs and
endwise does seem less annoying than the default lexima behaviour.

To be fair to lexima, I'm sure can be configured to be less annoying,
but auto-pairs and endwise combined seems to do the job for me.