Fix z command when sourced from bash profile
In bash, z appends a command to PROMPT_COMMAND to maintain its database,
so sourcing z must come after setting PROMPT_COMMAND in bash profile.

Prior to this, the z command existed, but the database was not being
maintained, which I didn't even notice as I had an existing database.