Stricter detection of yaml/json dotfiles in vimrc
I'd like to relax the rules re which dotfiles are examined to guess the
filetype so that it works for dotfiles outside the $HOME directory, but
the rules for filetype detection need to be stricter first. This is just
a start, be a little stricter about yaml and json filetype detection.