Fix use of did_filetype() in vimrc
Vim's built-in filetype detection does not run on BufWrite, so
did_filetype() will always return FALSE on BufWrite. The aim of the
initial change was to allow Vim's fallback filetype detection to be
overridden as it assumes if any of the first five lines start with #
then if must be a conf file. That particularly causes problems with
detection of yaml and ini files, both of which allow # for comments.

Fix just being explicit about overriding Vim's fallback detection of
conf files - if did_filetype() is FALSE and filetype is conf, assume
filetype was detected as FALLBACK, and allow override from vimrc.