Remove Homebrew bundle lockfile from repo
This is not used to pin to specific versions, just for comparing the
environment and versions installed between runs of `brew bundle`.

That means it's only really useful on the same machine as it generates
different output on different machines anyway, and I'm only using it to
get started on a new machine, so just remove the lockfile from the repo.