Fix visual mode cursor pos hack on AWS workspace
Can't safely map ESC on AWS workspace, so bring back ModeChanged autocmd
event, but abuse start and end position of changed text marks to allow
for deletion, without which we could jump to a position where text had
been deleted. Retain the ESC mapping for Mac as the behaviour with ESC
is slightly different, and better, after line deletion moves to first
word of next line, rather than first col of next line with ModeChanged.

It might be possible to tidy this up to remove the ESC mapping entirely,
and use the relevant mark motion, ` or ', to jump to correct position
after deletion depending on whether lines or characters were deleted,
but I'm not sure how to check that to choose to correct mark motion.