Don't automatically use pry as rails console
Pry is cool, I like it a lot, but I don't do much Rails development
these days, and when I do I'm usually helping someone else, in which
case I probably want to be using the same rails console as they are.

So, don't automatically start pry when running a rails console, just
make it available so I can use it when I need/want some of its features.