Hide eob fillchar when line numbers enabled in vim
Experimental, do we really need to show a tilde at the start of each
unused line in a buffer when line numbers are generally enabled? With
line numbers enabled it's easy to see unused lines at the end of a
buffer, they have no line number, and when nonumber has been set local
to a window, it's usually because that window is showing something you
don't want to edit, e.g. a diff, file explorer, tag outline view etc.