Workaround for vim-vinegar not keeping alt file
This is an issue with Vim, not Vinegar, but I haven't had time to
investigate in any detail so I'm just hacking a workaround for now.

Vinegar executes :edit %:h to open Netrw. With Vim 9.0.270 this retains
the alt file. With Vim 9.1.600 this sets the alt file to the directory
which was opened, i.e. expand('%:h'). Not sure when this changed or why.