Probably bat shit crazy NERDTree auto-reveal hacks
I don't use NERDTree very often, but when I am using it, it's normally
navigating a codebase I'm unfamiliar with, and automatically revealing
the current file in the tree is really handy (like VSCode does).

This will probably break, there's probably a damn good reason why this
was never included in NERDTree itself, I may well revert to a mapping.