Update gitgutter mappings, comment plugin clash
Comment plugin distributed with Vim9 now includes ac and ic text object
mappings for comment blocks, which clash with the gitgutter mappings. I
found the gitgutter mappings strange anyway, a mix of using h to refer
to hunks and c for, presumably, changes, so just use h for them all.